requirements.txt
streamlit
pillow
pytesseract
pyttsx3
opencv-python-headless
numpy